Drop selector_loop module; new JavaCoordinator does not need it
The rewritten ``JavaCoordinator`` (``airflow.sdk.coordinators.java``) lets
the JVM connect directly to two listening sockets (``comm`` and ``logs``)
and uses the accepted sockets as the supervisor's ``stdin`` / log pipes
straight up. There is no bytes-bridge between two sockets, so
``make_raw_forwarder`` -- the only helper the extracted ``selector_loop``
module added beyond what was already inline in ``supervisor.py`` -- has
no caller.

``_JavaActivitySubprocess`` reuses ``_register_pipe_readers`` and
``_close_unused_sockets`` via subclassing ``ActivitySubprocess``; both
methods existed before the extraction and remain in ``supervisor.py``
after this revert. The inline selector dispatch loop and
``make_buffered_socket_reader`` come back into ``supervisor.py`` so the
existing call sites (including the ``triggerer_job_runner`` re-export)
keep working unchanged.

This reverts commit 56464a8d45a96c8d15639784c0ca40367d22aae2 ("Add common
selector loop utilities for socket I/O handling for subprocesses") and
deletes ``test_selector_loop.py`` introduced by 4b80753e16.
